# swagger/src/main/resources/static/swagger-apis/user-service-api/swagger.yml
swagger: "2.0"
info:
    description: "Service handling operations associated with users"
    version: "1.0.0"
    title: "User service"
host: "user-service.taxi-microservice.com"
basePath: "/"
schemes:
    - "https"
    - "http"
tags:
    - name: "user"
      description: "Users related operations"
    - name: "user-settings"
      description: "User settings related operations"
    - name: "user-roles"
      description: "User roles related operations"
paths:
    /users:
        get:
            tags:
                - "user"
            summary: "Get list of users"
            description: ""
            operationId: "getUsers"
            produces:
                - "application/json"
            parameters:
                -   in: "query"
                    type: "integer"
                    name: "page"
                    required: true
                -   in: "query"
                    type: "integer"
                    name: "count"
                    required: true
            responses:
                "200":
                    description: "Users successfully returned"
                    schema:
                        type: "array"
                        items:
                            $ref: "#/definitions/UserResponseDTO"
                "204":
                    description: "No users"
                "401":
                    description: "Unauthorized user"
                "403":
                    description: "Access forbidden"
                "422":
                    description: "Invalid input"

    /users/{userId}:
        get:
            tags:
                - "user"
            summary: "Get user with given ID"
            description: ""
            operationId: "getUserWithId"
            produces:
                - "application/json"
            parameters:
                -   in: "path"
                    type: "string"
                    name: "userId"
                    required: true
            responses:
                "200":
                    description: "User with given ID successfully returned"
                    schema:
                        items:
                            $ref: "#/definitions/UserResponseDTO"
                "401":
                    description: "Unauthorized user"
                "403":
                    description: "Access forbidden"
                "404":
                    description: "User with given ID was not found"

        put:
            tags:
                - "user"
            summary: "Update user with given ID"
            description: ""
            operationId: "updateUser"
            consumes:
                - "application/json"
            produces:
                - "application/json"
            parameters:
                -   in: "path"
                    type: "string"
                    name: "userId"
                    required: true
                -   in: "body"
                    name: "body"
                    description: "User object that needs to be updated"
                    required: true
                    schema:
                        $ref: "#/definitions/UserUpdateDTO"
            responses:
                "200":
                    description: "User successfully updated"
                    schema:
                        items:
                            $ref: "#/definitions/UserUpdateDTO"
                "401":
                    description: "Unauthorized user"
                "403":
                    description: "Access forbidden"
                "404":
                    description: "User with given ID was not found"
        delete:
            tags:
                - "user"
            summary: "Delete user with given ID"
            description: ""
            operationId: "deleteUser"
            parameters:
                -   in: "path"
                    type: "string"
                    name: "userId"
                    required: true
            responses:
                "200":
                    description: "User successfully deleted"
                    schema:
                        items:
                            $ref: "#/definitions/UserResponseDTO"
                "401":
                    description: "Unauthorized user"
                "403":
                    description: "Access forbidden"
                "404":
                    description: "User with given ID was not found"
    /users/roles/{userId}:
        get:
            tags:
                - "user-roles"
            summary: "Get roles of user with given ID"
            description: ""
            produces:
                - "application/json"
            parameters:
                -   in: "path"
                    type: "string"
                    name: "userId"
                    required: true
            responses:
                "200":
                    description: "User roles successfully returned"
                    schema:
                        items:
                            type: "array"
                            $ref: "#/definitions/UserRoleDTO"
                "401":
                    description: "Unauthorized user"
                "403":
                    description: "Access forbidden"
                "404":
                    description: "User with given ID was not found"
        post:
            tags:
                - "user-roles"
            summary: "Add new role for user with given ID"
            description: "New roles can be added for user only by admin"
            produces:
                - "application/json"
            consumes:
                - "application/json"
            parameters:
                -   in: "path"
                    type: "string"
                    name: "userId"
                    required: true
                -   in: "body"
                    name: "body"
                    description: "User role object that needs to be added"
                    required: true
                    schema:
                        $ref: "#/definitions/UserRoleDTO"
            responses:
                "200":
                    description: "Successfully added role for user with given ID"
                    schema:
                        items:
                            $ref: "#/definitions/UserRoleDTO"
                "401":
                    description: "Unauthorized user"
                "403":
                    description: "Access forbidden"
                "404":
                    description: "User with given ID was not found"
        delete:
            tags:
                - "user-roles"
            summary: "Delete role with given id for user with given ID"
            description: "Roles can be deleted only by admin"
            produces:
                - "application/json"
            parameters:
                -   in: "path"
                    type: "string"
                    name: "userId"
                    required: true
                -   in: "query"
                    type: "string"
                    name: "roleId"
                    required: true
            responses:
                "200":
                    description: "Successfully deleted role of user with given ID"
                    schema:
                        items:
                            $ref: "#/definitions/UserRoleDTO"
                "401":
                    description: "Unauthorized user"
                "403":
                    description: "Access forbidden"
                "404":
                    description: "User or role with given ID was not found"

    /users/settings/{userId}:
        get:
            tags:
                - "user-settings"
            summary: "Get settings for user with given ID"
            description: ""
            operationId: "getUserSettings"
            produces:
                - "application/json"
            parameters:
                -   in: "path"
                    type: "string"
                    name: "userId"
                    required: true
            responses:
                "200":
                    description: "User settings successfully returned"
                    schema:
                        items:
                            $ref: "#/definitions/UserSettingsDTO"
                "401":
                    description: "Unauthorized user"
                "403":
                    description: "Access forbidden"
                "404":
                    description: "User with given ID was not found"
        put:
            tags:
                - "user-settings"
            summary: "Update settings for user with given ID"
            description: ""
            operationId: "updateUserSettings"
            consumes:
                - "application/json"
            produces:
                - "application/json"
            parameters:
                -   in: "path"
                    type: "string"
                    name: "userId"
                    required: true
                -   in: "body"
                    name: "body"
                    description: "User settings object that needs to be updated"
                    required: true
                    schema:
                        $ref: "#/definitions/UserSettingsDTO"
            responses:
                "200":
                    description: "User settings successfully returned"
                    schema:
                        items:
                            $ref: "#/definitions/UserSettingsDTO"
                "401":
                    description: "Unauthorized user"
                "403":
                    description: "Access forbidden"
                "404":
                    description: "User with given ID was not found"

    /current-user-data:
        get:
            tags:
                - "user"
            summary: "Return current user data"
            description: ""
            operationId: "getCurrentUserData"
            produces:
                - "application/json"
            responses:
                "200":
                    description: "Current user data successfully returned"
                    schema:
                        items:
                            $ref: "#/definitions/UserResponseDTO"
                "401":
                    description: "Unauthorized user"

definitions:
    UserUpdateDTO:
        type: "object"
        properties:
            name:
                type: "string"
            surname:
                type: "string"
            password:
                type: "string"
            pesel:
                type: "string"
            birthDate:
                type: "string"
                format: "date"

    UserResponseDTO:
        type: "object"
        properties:
            id:
                type: "integer"
                format: "long"
            userName:
                type: "string"
            email:
                type: "string"
            name:
                type: "string"
            surname:
                type: "string"
            pesel:
                type: "string"
            birthDate:
                type: "string"
                format: "date"
            userRole:
                $ref: "#/definitions/UserRole"

    UserSettingsDTO:
        type: "object"
        properties:
            appearance:
                $ref: "#/definitions/AppearanceType"
            language:
                type: "string"

    UserRoleDTO:
        type: "object"
        properties:
            role:
                type: "string"

    UserRole:
        type: "string"
        enum: ["driver", "passenger"]

    AppearanceType:
        type: "string"
        enum: ["light", "dark"]